Move AIoDClientWrapper to service/aiod
